requirements.txt
fastapi==0.110.3
uvicorn[standard]==0.30.1
streamlit==1.37.1
langchain==0.2.16
langchain-community==0.2.16
langchain-openai==0.1.23
langchain-text-splitters==0.2.2
faiss-cpu==1.13.2
rank-bm25==0.2.2
datasets==2.20.0
ragas==0.1.14
pandas==2.2.2
numpy==1.26.4
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
pydantic==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
httpx==0.27.0
boto3==1.34.149
tavily-python==0.4.0
pypdf==4.3.1
pdfplumber==0.11.9
pytest==8.2.2
pytest-asyncio==0.23.8
pytest-cov==5.0.0
